feat(client): add environment parameter to client initialization
This adds a new ENVIRONMENT enum to the SDK API, that may be used to configure the client for use
with the Truework Sandbox. If this argument is passed, baseURL MUST NOT be passed to avoid ambiguous
configurations. This constraint is enforced by throwing an exception if an invalid configuration is
used. The client defaults to the production API for backwards compatibility.

API-46