fix(client): use latest version if no version is defined
The behavior of the SDK was not aligned with the documentation: the documentation states that the
latest API version will be used if no version is passed, but the version was actually being
hardcoded to 2019-10-15.

BREAKING CHANGE: If no version parameter is passed during client initialization, the latest API
version will now be used instead of 2019-10-15.